- Added Role, User model. - Changed generic user features to UserController - Removed field discipline_id from signs

+++ app/Http/Controllers/StudentController.php
@@ -4,7 +4,8 @@ namespace App\Http\Controllers;
 
 use App\Course;
 use App\Discipline;
-use App\Student;
+use App\Role;
+use App\User;
 use Illuminate\Http\Request;
 
 class StudentController extends Controller
@@ -12,15 +13,15 @@ class StudentController extends Controller
     /**
      * @var
      */
-    protected $student;
+    protected $user;
 
     /**
      * StudentController constructor.
-     * @param Student $student
+     * @param User $user
      */
-    public function __construct(Student $student)
+    public function __construct(User $user)
     {
-        $this->student = $student;
+        $this->user = $user;
     }
 
     /**
@@ -28,7 +29,7 @@ class StudentController extends Controller
      */
     public function index()
     {
-        $students = $this->student->all();
+        $students = $this->user->allStudents()->get();
 
         return response()->json($students);
     }
@@ -39,8 +40,14 @@ class StudentController extends Controller
      */
     public function getStudent($id)
     {
+        $student = $this->user->find($id);
 
-        $student = $this->student->find($id);
+        if ($student == null || ! $student->isStudent()) {
+            return response()->json([
+                'status'  => 400,
+                'message' => 'The Student ID provided is not a student or not found!'
+            ]);
+        }
 
         return response()->json($student);
     }
@@ -51,7 +58,11 @@ class StudentController extends Controller
      */
     public function createStudent(Request $request)
     {
-        $student = $this->student->create($request->all());
+        $student = $this->user->create($request->all());
+
+        $student->roles()->create([
+           'role' => Role::STUDENT
+        ]);
 
         return response()->json($student);
     }
@@ -62,8 +73,19 @@ class StudentController extends Controller
      */
     public function deleteStudent($id)
     {
-        $student = $this->student->find($id);
+        $student = $this->user->find($id);
+
+        if ($student == null || ! $student->isStudent()) {
+            return response()->json([
+                'status'  => 400,
+                'message' => 'The Student ID provided is not a student or not found!'
+            ]);
+        }
+
         $student->delete();
+
+        $role = Role::where('user_id',$student->id)->where('role', Role::STUDENT);
+        $role->delete();
 
         return response()->json([
             'status'  => 500,
@@ -72,33 +94,19 @@ class StudentController extends Controller
     }
 
     /**
-     * @param Request $request
+     * Gets the grades from a Student
+     *
      * @param $id
      * @return \Illuminate\Http\JsonResponse
      */
-    public function updateUser(Request $request, $id)
-    {
-        $student = $this->student->find($id);
-        $student->name = $request->input('name');
-        $student->surname = $request->input('surname');
-        $student->ssn = $request->input('ssn');
-        $student->birthdate = $request->input('birthdate');
-        $student->gpa = $request->input('gpa');
-        $student->gender = $request->input('gender');
-        $student->save();
-
-        return response()->json($student);
-    }
-
     public function getGrades($id)
     {
-        $student = Student::find($id);
+        $student = User::find($id);
 
-        if($student == null)
-        {
+        if ($student == null || ! $student->isStudent()) {
             return response()->json([
-                'status'  => 404,
-                'message' => 'Student not found!'
+                'status'  => 400,
+                'message' => 'The Student ID provided is not a student or not found!'
             ]);
         }
 
@@ -107,8 +115,61 @@ class StudentController extends Controller
         return response()->json($grades);
     }
 
-    public function enrollCourse(Request $request)
+    /**
+     * Gets enrolled courses from a Student
+     *
+     * @param $id
+     * @return \Illuminate\Http\JsonResponse
+     */
+    public function getCourses($id)
     {
+        $student = User::find($id);
+
+        if ($student == null || ! $student->isStudent()) {
+            return response()->json([
+                'status'  => 400,
+                'message' => 'The Student ID provided is not a student or not found!'
+            ]);
+        }
+
+        $courses = $student->courses;
+
+        return response()->json($courses);
+    }
+
+    /**
+     * Gets all Classes from a Student
+     *
+     * @param $id
+     * @return \Illuminate\Http\JsonResponse
+     */
+    public function getClasses($id)
+    {
+        $student = User::find($id);
+
+        if ($student == null || ! $student->isStudent()) {
+            return response()->json([
+                'status'  => 400,
+                'message' => 'The Student ID provided is not a student or not found!'
+            ]);
+        }
+
+        $classes = $student->classes;
+
+        return response()->json($classes);
+    }
+
+    /**
+     * Enroll a student in a desired course
+     *
+     * @param Request $request
+     * @param null $id
+     * @return \Illuminate\Http\JsonResponse
+     */
+    public function enrollCourse(Request $request, $id = null)
+    {
+        //TODO: Adicionar middleware para verificar permissoes
+
         $course = Course::find($request->input('course_id'));
 
         if ($course == null) {
@@ -118,17 +179,16 @@ class StudentController extends Controller
             ]);
         }
 
-        $student = Student::where('token',$request->input('token'))->first();
+        $student = User::find($id);
 
-        if ($student == null) {
+        if ($student == null || ! $student->isStudent()) {
             return response()->json([
-                'status'  => 404,
-                'message' => 'Token is invalid!'
+                'status'  => 400,
+                'message' => 'The Student ID provided is not a student or not found!'
             ]);
         }
 
-        if($student->courses->has($course->id))
-        {
+        if ($student->courses->has($course->id)) {
             $student->courses()->detach($course->id);
 
             return response()->json([
@@ -145,7 +205,14 @@ class StudentController extends Controller
         }
     }
 
-    public function enrollDiscipline(Request $request)
+    /**
+     * Enroll a student on a desired discipline
+     *
+     * @param Request $request
+     * @param $id
+     * @return \Illuminate\Http\JsonResponse
+     */
+    public function enrollDiscipline(Request $request, $id)
     {
         $course = Discipline::find($request->input('discipline_id'));
 
@@ -156,7 +223,7 @@ class StudentController extends Controller
             ]);
         }
 
-        $student = Student::where('token',$request->input('token'))->first();
+        $student = User::find($id);
 
         if ($student == null) {
             return response()->json([
@@ -164,5 +231,27 @@ class StudentController extends Controller
                 'message' => 'Token is invalid!'
             ]);
         }
+
+        if ($student->courses->has($course->id)) {
+            $student->courses()->detach($course->id);
+
+            return response()->json([
+                'status'  => 500,
+                'message' => 'Student ' . $student->id . ' has been removed from course ' . $course->id,
+            ]);
+        } else {
+            $student->courses()->attach($course->id);
+
+            return response()->json([
+                'status'  => 500,
+                'message' => 'Student ' . $student->id . ' has added to course ' . $course->id,
+            ]);
+        }
+    }
+
+
+    public function test()
+    {
+        return response()->json(\App\Discipline::find(1)->classes);
     }
 }
